add formula 6.47 and its sub-factors to implementation
